Enable addTodo
Now it's time to add an event handler. An event is (in general) triggered by the user of a website. e.g. clicking or using the keyboard triggers events.

We currently need an event handler, that adds a new item to the list of todos when the enter key on the keyboard is pressed.

In Vue, you bind event handlers in the template with an @ (e.g. @click=""). Here we add @keyup.enter="addTodo" which means, that the "addTodo" method will be called, when the user presses the "enter" key on the keyboard.

The "addTodo" then takes the value of the input field and creates a new todo entry, before resetting the input field.