Add remove functionality
Now that we can add new todo items, let's add the removal of an item. This is also done by an event handler.

Each item has a button with the class "destroy". We add a "click" event handler to this button. This time we also give the current "todo" item to the executed method. Only this way we are able to remove the correct todo item from the list of todos.

The "removeTodo" method then searches for the given todo inside the list of todos and removes it (splicing).