Enable state handling and styling
Until now, the toggling of the checkbox had no effect. Even though a check mark is shown, when clicking in the circle, the state of the item is actually not updated.

We can test this, by adding a dynamic "complete" class to the list element. Since we want the class to represent the current item state, we will use Vues attribute binding: :attribute="value"

In case of the "class" attribute we can use a "JavaScript Object" to let Vue decide, if the "completed" class needs to be added to the element or not depending on the value of "isDone" (which is "true" or "false").

After you added ":class="{ completed: todo.isDone }"" you will notice, that the styling of the item does still not change, when you check/uncheck it. This is because the "isDone" value of the todo is never changed.

To make this work, we need to tell Vue, that it should update the "isDone" value of the clicked item. This can be simply done by adding "v-model="todo.isDone"" to the "toggle" checkbox.