Make item counter dynamic
Until now, the item counter in the lower left was also static. Let's change this.

The counter is defined in the "TodoListFooter" component. The component is then imported in the "TodoList" component and displayed.

In Vue a component can give any kind of data to a child component with "props". Just like "data" on "TodoList", "props" add new fields to the component, which then can be used in the template.

To make the item counter dynamic, you need to define a new prop "amountOfItems" on the "TodoListFooter" and replace the static value "5" in the template with an interpolation ("{{ amountOfItems }}). If you now hand the current amount of items from "TodoList" to "TodoListFooter" with an attribute binding (:amountOfItems="todos.length"), the counter will be dynamic.