Update item counter
Even though the counter is now interactive, it counts the amount of items, rather then the left items.

To change this we need something new: a "computed". A "computed" is a calculated which is based on other values in the component. In out case we want to count the amount of todo items which don't have a "isDone" value of "true".

Once the "computed" is created, we update to original value binding in the template with the new computed value.