Enabling global state toggling
On the left side of the input field for new entries, there is a toggle to set all todos "isDone" to "true" or "false".

We can also use a computed to introduce this behaviour. But this time the computed is not just a function like "remaining" but has a "get" and "set" method. This way we can do two different actions, based on the fact if the value of the computed is requested (get) or changed (set).

In the "get" we check if there are any items left, that still have a "isDone" status with "false". This is needed to determine the state and styling of the toggle.

In the "set" we modify the "isDone" value of each todo item and therefore check/uncheck them all.