Enabling filtering
Lets enable some more functionality: the filtering  of item state.

For this we need something new: "custom events". Custom events are events that you can define yourself and are needed to send data from a child component to the parent component. In this case we "emit" a "filter" event, which holds the name of the clicked filter.

On the parent we need to also change the rendering of the todo list. Currently we render the whole list of todo items. When we want to use a filter, we need to define a computed ("filteredItems") which will only return the todo items, which meet the current filter.

After also defining an event handler for the custom "filter" event. It saves the given filter, which itself will be used to calculate the list of "filteredItems".