Enabling "clear completed"
Time for the last step, to complete our "application": auto removal of all finished entries.

To do this we first need a new event handler @click on the footer component. This times we "emit" the event directly from the template.

On the parent component we will remove the done items from the todo list when the event arrives and only keep the items, that are not done yet.